Recurrent neural network
Implements a recurrent neural network to analyze
the Kato neuronal timeseries data. Includes rnn.py,
a module to access the RNN, and two notebooks
demonstrating possible applications of the RNN.

Also included in this commit is a notebook (kato_visualization)
that was part of the old repo hosted by theideasmith. It provides
some examples of processing and visualizing the Kato data.