Numerical differentiation and Improvements to RNN
- Implements numerical differentiation (Chartrand, et al.)
as cited in Kato, using off-the-shelf script by Chartrand.
Numerical differentiation is available as a Analyzer method 'tvd'

- Demonstrates the use of numerical differentiation in a Jupyter
notebook (KatoPipeline.ipynb). The notebook compares the results
of PCA on Kato's provided derivative data vs computed derivative data

- Implements Analyzer methods to quickly generate 'Kato-style'
PCA plots and timeseries plots

- RNN: based on feedback from @jrieke, attempts to implement
GMMActivation and GMM Loss functions on RNN. However, the code
is not currently working correctly, and is returning only NaNs
when attempting to fit the model.